Extract Threat from ScanStateResponse & introduce ThreatResponse
