+++ fluxc/src/main/java/org/wordpress/android/fluxc/network/rest/wpcom/scan/ScanStateResponse.kt
@@ -2,6 +2,7 @@ package org.wordpress.android.fluxc.network.rest.wpcom.scan
 
 import com.google.gson.annotations.SerializedName
 import org.wordpress.android.fluxc.network.Response
+import org.wordpress.android.fluxc.network.rest.wpcom.scan.threat.Threat
 import java.util.Date
 
 data class ScanStateResponse(
@@ -13,30 +14,6 @@ data class ScanStateResponse(
     @SerializedName("most_recent") val mostRecentStatus: ScanProgressStatus?,
     @SerializedName("current") val currentStatus: ScanProgressStatus?
 ) : Response {
-    data class Threat(
-        @SerializedName("id") val id: Int,
-        @SerializedName("signature") val signature: String,
-        @SerializedName("description") val description: String,
-        @SerializedName("status") val status: String,
-        @SerializedName("fixable") val fixable: Fixable?,
-        @SerializedName("extension") val extension: Extension?,
-        @SerializedName("first_detected") val firstDetected: Date,
-        @SerializedName("fixed_on") val fixedOn: Date?
-    ) {
-        data class Fixable(
-            @SerializedName("fixer") val fixer: String?,
-            @SerializedName("target") val target: String?
-        )
-
-        data class Extension(
-            @SerializedName("type") val type: String?,
-            @SerializedName("slug") val slug: String?,
-            @SerializedName("name") val name: String?,
-            @SerializedName("version") val version: String?,
-            @SerializedName("isPremium") val isPremium: Boolean?
-        )
-    }
-
     data class Credentials(
         @SerializedName("type") val type: String,
         @SerializedName("role") val role: String,
