+++ fluxc/src/main/java/org/wordpress/android/fluxc/network/rest/wpcom/scan/threat/ThreatResponse.kt
@@ -1,0 +1,9 @@
+package org.wordpress.android.fluxc.network.rest.wpcom.scan.threat
+
+import com.google.gson.annotations.SerializedName
+import org.wordpress.android.fluxc.network.Response
+
+data class ThreatResponse(
+    @SerializedName("success") val success: Boolean?,
+    @SerializedName("threat") val threat: Threat? = null
+) : Response
